// src/store/modules/App/enums.ts
export  enum TypeNames {
  HANDLE_ADD_EVENT_TO_PAKEEP = 'HANDLE_ADD_EVENT_TO_PAKEEP',
  HANDLE_ADD_NEW_PAKEEP = 'HANDLE_ADD_NEW_PAKEEP',
  HANDLE_CHANGE_MENU_OPEN_STATUS = 'HANDLE_CHANGE_MENU_OPEN_STATUS',
  HANDLE_DELETE_PAKEEP = 'DELETE_PAKEEP',
  HANDLE_SET_NEW_ORDER_NAMES = 'SET_NEW_ORDER_NAMES',
  HANDLE_SET_CURRENT_FOLDER_PROPERTY_IDX = 'HANDLE_CURRENT_FOLDER_PROPERTY_IDX',
  HANDLE_CHANGE_FOLDERS = 'HANDLE_CHANGE_FOLDERS',
  HANDLE_CHANGE_GLOBAL_LABELS = 'HANDLE_CHANGE_GLOBAL_LABELS',
  HANDLE_CHANGE_LABELS_IN_PAKEEP = 'HANDLE_CHANGE_LABELS_IN_PAKEEP',
  HANDLE_SET_DRAWER_WIDTH = 'HANDLE_SET_DRAWER_WIDTH',
  HANDLE_ADD_NEW_GLOBAL_LABEL = 'HANDLE_ADD_NEW_GLOBAL_LABEL',
  HANDLE_CHANGE_PAKEEPS = 'HANDLE_CHANGE_PAKEEPS',
  HANDLE_PIN_STATUS_OF_PAKEEPS = 'HANDLE_PIN_STATUS_OF_PAKEEPS',
  HANDLE_SET_ORDER_NAMES_OF_PINNED_PAKEEPS = 'HANDLE_SET_ORDER_NAMES_OF_PINNED_PAKEEPS',
  HANDLE_SET_SELECTED_PAKEEPIDS_ARR = 'HANDLE_SET_SELECTED_PAKEEPIDS_ARR',
  HANDLE_CANCEL_SELECTING_STATUS = 'HANDLE_CANCEL_SELECTING_STATUS',
  HANDLE_CHANGE_SELECTED_PAKEEPS_PROPERTY = 'HANDLE_CHANGE_SELECTED_PAKEEPS_PROPERTY',
  HANDLE_ADD_LABEL_TO_PAKEEP = 'ADD_LABEL_TO_PAKEEP',
  HANDLE_CHANGE_PAKEEP_PROPERTY = 'HANDLE_CHANGE_PAKEEP_PROPERTY',
  HANDLE_CHANGE_THEME_COLORS = 'HANDLE_CHANGE_THEME_COLORS'
} 